Major update. Create Instrument Mode and Chromatic Mode, with animations between each mode. Refactor TunerViewModel and Tuner to support the two modes.

+++ app/src/main/java/tech/ajsf/instrutune/features/tuner/view/TunerMeterFunctions.kt
@@ -1,6 +1,6 @@
 package tech.ajsf.instrutune.features.tuner.view
 
-fun scaleAndCurveDelta(delta: Float): Int {
+fun scaleAndCurveDelta(delta: Int): Int {
     val absDelta = Math.abs(delta.toDouble())
     val intDelta = Math.pow(absDelta, .63).toInt()
     return Math.min(intDelta, 16)
